move Network loggers to separate files
