+++ types/relay-runtime/lib/network/RelayNetworkLoggerTransaction.d.ts
@@ -1,0 +1,30 @@
+import { RequestParameters } from '../util/RelayConcreteNode';
+import { Variables, CacheConfig, UploadableMap, GraphQLResponse } from '../../index';
+
+export interface RelayNetworkLog {
+    label: string;
+    values: ReadonlyArray<any>;
+}
+
+export interface LoggerTransactionConfig {
+    request: RequestParameters;
+    variables: Variables;
+    cacheConfig: CacheConfig;
+    uploadables?: UploadableMap;
+}
+
+export class RelayNetworkLoggerTransaction {
+    constructor(config: LoggerTransactionConfig);
+    addLog(label: string, ...values: any[]): void;
+    commitLogs(error: Error, payload: GraphQLResponse, status?: string): void;
+    flushLogs(error: Error, payload: GraphQLResponse, status?: string): void;
+    markCommitted(): void;
+    getCacheConfig(): CacheConfig;
+    getIdentifier(): string;
+    getLogsToPrint(): RelayNetworkLog[];
+    getRequest(): RequestParameters;
+    getUploadables(): UploadableMap;
+    getVariables(): Variables;
+    timerStart(): void;
+    timerEnd(): void;
+}
